requirements.txt
streamlit==1.34.0
python-dotenv==1.0.1
openai==1.30.1
langchain==0.1.14
langchain-openai==0.0.6